# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: AniMeat API
  description: |
    Server-authoritative API for the AniMeat food delivery platform.
    Source of truth for money, orders, payments, roles, and state transitions.
    All timestamps are RFC3339 UTC. Money is transmitted as integer paise (minor units)
    where a `*_paise` field exists; legacy rupee fields may still be present for
    backwards compatibility with legacy Flutter clients.
  version: 1.0.0
  contact:
    name: AniMeat Engineering
servers:
  - url: https://api.animeat.app
    description: Production
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/v1/users/me:
    get:
      summary: Get own profile
      operationId: getProfile
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404': { $ref: '#/components/responses/Error' }
    put:
      summary: Upsert profile (created on first login)
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/v1/users/me/notifications:
    get:
      summary: List own notifications
      operationId: listNotifications
      parameters:
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error, code, message]
      properties:
        error: { type: string }
        code: { type: string }
        message: { type: string }
        detail: { type: string }
...
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        address: { type: string }
        avatar_url: { type: string }
    UpdateProfileRequest:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        address: { type: string }
        avatar_url: { type: string }
    Notification:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        body: { type: string }
        notification_type: { type: string }
        entity_type: { type: string }
        entity_id: { type: string }
        is_read: { type: boolean }
        created_at: { type: string, format: date-time }
...